+++ docs/NEWS.md
@@ -1,5 +1,6 @@
 ### UNRELEASED
 
+ * *feature:* new **--format** string **%{progress-conemu}** to show progress in the terminal title tab on terminals with ConEmu OSC 9;4 support ([#202](https://codeberg.org/ivarch/pv/issues/202))
  * *cleanup:* eliminate potential for closure of standard output under a combination of error conditions when using **--monitor** mode ([#201](https://codeberg.org/ivarch/pv/pulls/201))
 
 ### 1.11.0 - 11 June 2026